bio_hama_project/requirements.txt
# Bio-HAMA Project Dependencies
# Core ML Framework (PyTorch will be installed separately for GPU)
# Install PyTorch with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Hugging Face Transformers
transformers>=4.30.0

# Data Handling and Numerics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.6.1

# Additional utilities
tabulate>=0.9.0

